Simplify Segment to a single Index-based indexer (#84)
Replace the three public indexers (long, int, Index) with a single
this[Index] backed by a private ItemAt(long) method. int implicitly
converts to Index, so segment[0], segment[i], and segment[^1] all
resolve without ambiguity — the int overload added for disambiguation
is no longer needed.

+++ tests/Fabrica.Core.Tests/Collections/ProducerConsumerQueueSingleThreadStressTests.cs
@@ -76,7 +76,7 @@ public class ProducerConsumerQueueSingleThreadStressTests
             var segment = queue.ConsumerAcquire();
             Assert.Equal(totalProduced - totalConsumed, segment.Count);
 
-            for (var i = 0L; i < segment.Count; i++)
+            for (var i = 0; i < (int)segment.Count; i++)
                 Assert.Equal(totalConsumed + i, segment[i]);
 
             totalConsumed += segment.Count;
